fixed compiler warning

+++ boot/boot.cpp
@@ -62,7 +62,7 @@ namespace KrollBoot
 		FindUpdate();
 	
 		vector<SharedDependency> missing = app->ResolveDependencies();
-		for (int i = 0; i < missing.size(); i++)
+		for (size_t i = 0; i < missing.size(); i++)
 		{
 			SharedDependency d = missing.at(i);
 			std::cerr << "Unresolved: " << d->name << " " << d->version << std::endl;
